requirements.txt
Django<1.9
psycopg2
ipython
django-reversion
django-model-utils
django-grappelli
